Refactor e2e tally logic into a dedicated step
