+++ src/electionguard_cli/steps/import_ballots/import_ballots_input_retrieval_step.py
@@ -1,4 +1,3 @@
-from contextvars import Context
 from io import TextIOWrapper
 from os import listdir
 from os.path import join
